added sun motion
