+++ App.js
@@ -4,7 +4,6 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import LoginScreen from "./screens/LoginScreen";
 import QrCodeScannerScreen from "./screens/QrScannerScreen";
 import CommunityScreen from "./screens/CommunityScreen";
-import LocationTestScreen from "./screens/LocationTestScreen";
 import NavBar from "./components/NavBar";
 import { LogBox } from "react-native";
 
@@ -25,7 +24,6 @@ export default App = () => {
         <Stack.Screen name="App" component={NavBar} />
         <Stack.Screen name="QrScanner" component={QrCodeScannerScreen} />
         <Stack.Screen name="Community" component={CommunityScreen} />
-        <Stack.Screen name="LocationTest" component={LocationTestScreen} />
       </Stack.Navigator>
     </NavigationContainer>
   );
